split qwen36 moe logits helpers
