+++ crates/runner/src/lib.rs
@@ -19,6 +19,7 @@ pub mod gemma4_int4_engine;
 pub mod oracle;
 pub mod prefill_engine;
 pub mod qwen36_moe_decode;
+pub mod qwen36_moe_logits;
 pub mod qwen36_moe_mtp;
 pub mod qwen36_moe_persistent_decode;
 pub mod qwen36_moe_residency;
