+++ crates/runner/src/qwen36_moe_decode.rs
@@ -1,12 +1,12 @@
-//! Host-orchestrated multi-launch decode for Qwen3.6-MoE — PR 4c step 2.
+//! Host-orchestrated multi-launch decode for Qwen3.6-MoE.
 //!
 //! Walks the hybrid pattern (every 4th layer full-attn at indices 3/7/11/...,
 //! every other layer linear-attn) calling the per-block FFI launchers
 //! ([`kernel_ffi::qwen36_moe::attn_step_launch`] / `linear_step_launch` /
 //! `ffn_step_launch`) at stage 5 (full-layer output). One HIP launch per
-//! block per layer per token — high launch overhead, but correct,
-//! reviewable, and unblocks end-to-end testing. The persistent megakernel
-//! (PR 4c step 4) folds these N×3 launches down to 1.
+//! block per layer per token. The persistent megakernel is the production
+//! path, but this chained implementation remains the reviewable oracle path
+//! for parity tests, fallback runs, and isolated stage diagnostics.
 //!
 //! The decode core in [`run_chained_decode`] takes pre-allocated
 //! per-layer weight + state buffers and an initial hidden vector, runs the
@@ -15,9 +15,8 @@
 //! `crates/runner/tests/qwen36_moe_multilayer_parity.rs` to gate
 //! correctness against the multi-layer Python oracle).
 //!
-//! [`host_final_norm_lm_head`] applies the final RMSnorm + lm_head GEMV on
-//! the host. The plan recommends host-side for PR 4c (~1ms slower per token
-//! vs a kernel; lifting it to GPU is PR 4d).
+//! Host-side final RMSnorm, lm_head helpers, INT4 dequant, and sampling live
+//! in `qwen36_moe_logits`.
 //!
 //! Both the parity test (synthetic weights from oracle JSON) and the
 //! engine's real-decode path (weights from the bake) call into the same
@@ -37,7 +36,7 @@ use kernel_ffi::qwen36_moe::{
     Qwen36MoeLinearStepParams, Qwen36MoeLinearStepWeights,
 };
 
-use crate::tensor_bytes;
+use crate::qwen36_moe_logits::bf16_bytes_to_f32;
 
 /// Hybrid pattern: every 4th layer is full attention. Indices 3, 7, 11, ...
 /// are full; everything else is linear. Matches Qwen3.6-MoE 35B-A3B.
@@ -1240,327 +1239,6 @@ fn download_topk_routes(
         .collect())
 }
 
-// ---------------------------------------------------------------------------
-// Host-side final RMSnorm + lm_head GEMV. The plan recommends host execution
-// for PR 4c; lifting to a dedicated kernel is PR 4d.
-// ---------------------------------------------------------------------------
-
-/// Decode a stream of BF16 little-endian bytes into F32. The oracle stores
-/// BF16 as raw int16 → bytes; this is the inverse.
-pub fn bf16_bytes_to_f32(bytes: &[u8]) -> Vec<f32> {
-    assert!(bytes.len() % 2 == 0, "BF16 bytes must be even");
-    tensor_bytes::bf16_bytes_to_f32(bytes)
-}
-
-/// Round an F32 to BF16 (RNE), returning the 16 raw bits. Matches PyTorch's
-/// `.to(torch.bfloat16)` rounding: nearest-even on the lopped-off mantissa
-/// bits, with the standard NaN-quieting convention.
-pub fn f32_to_bf16_bits(x: f32) -> u16 {
-    let bits = x.to_bits();
-    if (bits & 0x7FFF_FFFF) > 0x7F80_0000 {
-        // NaN: keep top 16 with quiet bit set.
-        return ((bits >> 16) | 0x0040) as u16;
-    }
-    let lsb = (bits >> 16) & 1;
-    let rounding_bias = 0x7FFFu32 + lsb;
-    let rounded = bits.wrapping_add(rounding_bias);
-    (rounded >> 16) as u16
-}
-
-/// Encode a slice of F32 values to BF16 little-endian bytes (RNE).
-pub fn f32_to_bf16_bytes(vals: &[f32]) -> Vec<u8> {
-    tensor_bytes::f32_to_bf16_bytes(vals)
-}
-
-/// Apply RMSnorm + lm_head GEMV on the host. Mirrors the multi-layer
-/// oracle's tail:
-///   final_normed = rms_norm(hidden, final_norm_w, eps)   # (1+w) offset
-///   logits       = final_normed.to(F32) @ lm_head_w.to(F32).T
-///   logits       = logits.to(BF16)
-///
-/// The RMSnorm uses the HuggingFace `Qwen3_5MoeRMSNorm` convention with
-/// the `(1.0 + weight)` unit offset — `model.norm` in the HF text model
-/// (line 1354 of `transformers/models/qwen3_5_moe/modeling_qwen3_5_moe.py`)
-/// is an instance of `Qwen3_5MoeRMSNorm` whose forward computes
-/// `output * (1.0 + self.weight.float())`.
-///
-/// All inputs are BF16 little-endian byte streams; output is BF16
-/// little-endian bytes for `vocab` logit channels.
-pub fn host_final_norm_lm_head(
-    hidden_bytes: &[u8],
-    final_norm_w_bytes: &[u8],
-    lm_head_w_bytes: &[u8],
-    hidden: usize,
-    vocab: usize,
-    eps: f32,
-) -> Vec<u8> {
-    assert_eq!(hidden_bytes.len(), hidden * 2, "hidden bytes mismatch");
-    assert_eq!(
-        final_norm_w_bytes.len(),
-        hidden * 2,
-        "norm_w bytes mismatch"
-    );
-    assert_eq!(
-        lm_head_w_bytes.len(),
-        vocab * hidden * 2,
-        "lm_head bytes mismatch"
-    );
-
-    // BF16-input convenience wrapper. For multi-token decode loops use
-    // `host_final_norm_lm_head_f32` directly with cached F32 lm_head + norm
-    // weight to avoid re-converting the (multi-GiB) lm_head matrix per step.
-    let w_f32 = bf16_bytes_to_f32(final_norm_w_bytes);
-    let lm_f32 = bf16_bytes_to_f32(lm_head_w_bytes);
-    host_final_norm_lm_head_f32(hidden_bytes, &w_f32, &lm_f32, hidden, vocab, eps)
-}
-
-/// Same math as [`host_final_norm_lm_head`] but with `final_norm_w` and
-/// `lm_head_w` already converted to F32 by the caller. Hot-loop variant —
-/// the BF16→F32 conversion of the lm_head matrix is by far the dominant
-/// cost on the 35B-A3B geometry (~1 GiB BF16 → ~2 GiB F32 per call), and
-/// it never changes across tokens, so the engine converts it once and
-/// passes the F32 view here per-step.
-pub fn host_final_norm_lm_head_f32(
-    hidden_bytes: &[u8],
-    final_norm_w_f32: &[f32],
-    lm_head_w_f32: &[f32],
-    hidden: usize,
-    vocab: usize,
-    eps: f32,
-) -> Vec<u8> {
-    assert_eq!(hidden_bytes.len(), hidden * 2, "hidden bytes mismatch");
-    assert_eq!(final_norm_w_f32.len(), hidden, "norm_w f32 len mismatch");
-    assert_eq!(
-        lm_head_w_f32.len(),
-        vocab * hidden,
-        "lm_head f32 len mismatch"
-    );
-
-    let h_f32 = bf16_bytes_to_f32(hidden_bytes);
-
-    // RMSnorm with the HF `Qwen3_5MoeRMSNorm` convention: F32 mean of
-    // squares -> rsqrt(var+eps) -> elementwise mul by `(1.0 + w)`. The
-    // stored weight is a small delta around zero (initialized to zeros
-    // in HF — line 810 of modeling_qwen3_5_moe.py) and the effective
-    // scale factor is `1 + w`.
-    let mean_sq: f32 = h_f32.iter().map(|&x| x * x).sum::<f32>() / hidden as f32;
-    let rsqrt = 1.0 / (mean_sq + eps).sqrt();
-    let normed: Vec<f32> = h_f32
-        .iter()
-        .zip(final_norm_w_f32.iter())
-        .map(|(&x, &w)| x * rsqrt * (1.0 + w))
-        .collect();
-
-    // GEMV: lm_head [vocab, hidden] @ normed [hidden] -> logits [vocab].
-    // F64 accumulator matches the oracle's torch.float32 reduction order
-    // closely enough for cos_sim parity at vocab+hidden ≤ a few thousand.
-    let mut logits = vec![0f32; vocab];
-    for v in 0..vocab {
-        let row_start = v * hidden;
-        let mut acc = 0f64;
-        for h in 0..hidden {
-            acc += lm_head_w_f32[row_start + h] as f64 * normed[h] as f64;
-        }
-        logits[v] = acc as f32;
-    }
-
-    f32_to_bf16_bytes(&logits)
-}
-
-/// Dequantize an INT4-packed weight slab back to BF16 bytes. Mirrors the
-/// kernel's `int4_dequant_scalar`: for each output row + input column the
-/// nibble is split off the byte (even col → low nibble, odd col → high),
-/// converted to F32, then `bf16(q*s - z*s)` written using the matching
-/// `[out/gs, in/gs]` BF16 scale/zero tile.
-///
-/// Used by the host-side lm_head path when the bake quantizes lm_head.
-/// `out_dim` × `in_dim` BF16 result = ~1 GiB for 35B-A3B (vocab 248K ×
-/// hidden 2048); fine for one-token smoke, kernel-side lm_head is PR 4d.
-pub fn dequant_int4_to_bf16_bytes(
-    packed: &[u8],
-    scale_bf16: &[u8],
-    zero_bf16: &[u8],
-    out_dim: usize,
-    in_dim: usize,
-    group_size: usize,
-) -> Vec<u8> {
-    assert_eq!(packed.len(), out_dim * in_dim / 2, "packed size mismatch");
-    assert_eq!(
-        in_dim % group_size,
-        0,
-        "in_dim must be divisible by group_size"
-    );
-    assert_eq!(
-        out_dim % group_size,
-        0,
-        "out_dim must be divisible by group_size"
-    );
-    let n_row_tiles = out_dim / group_size;
-    let n_col_tiles = in_dim / group_size;
-    assert_eq!(
-        scale_bf16.len(),
-        n_row_tiles * n_col_tiles * 2,
-        "scale size mismatch"
-    );
-    assert_eq!(
-        zero_bf16.len(),
-        n_row_tiles * n_col_tiles * 2,
-        "zero size mismatch"
-    );
-
-    let scale = bf16_bytes_to_f32(scale_bf16);
-    let zero = bf16_bytes_to_f32(zero_bf16);
-    let mut out = Vec::with_capacity(out_dim * in_dim * 2);
-    let half_in = in_dim / 2;
-    for o in 0..out_dim {
-        let row_tile = o / group_size;
-        let row_base = o * half_in;
-        for i in 0..in_dim {
-            let col_tile = i / group_size;
-            let tile_idx = row_tile * n_col_tiles + col_tile;
-            let s = scale[tile_idx];
-            let z = zero[tile_idx];
-            let byte = packed[row_base + (i / 2)];
-            let nib = if i % 2 == 0 {
-                byte & 0x0F
-            } else {
-                (byte >> 4) & 0x0F
-            };
-            let q = nib as f32;
-            // Single-rounding bf16(q*s - z*s) — matches the kernel's
-            // `bf16_round_rne_f32_finite(n*s - zs)`.
-            let bf = f32_to_bf16_bits(q * s - z * s);
-            out.push((bf & 0xFF) as u8);
-            out.push((bf >> 8) as u8);
-        }
-    }
-    out
-}
-
-/// Tiny dependency-free xorshift64 RNG. Deterministic given the seed —
-/// the engine's `--sampling-seed` plus identical prompt + model + sampling
-/// params produces bit-identical generation.
-pub struct XorshiftRng(u64);
-
-impl XorshiftRng {
-    pub fn new(seed: u64) -> Self {
-        // Xorshift requires a non-zero state.
-        Self(if seed == 0 {
-            0x9E37_79B9_7F4A_7C15
-        } else {
-            seed
-        })
-    }
-    pub fn next_u64(&mut self) -> u64 {
-        self.0 ^= self.0 << 13;
-        self.0 ^= self.0 >> 7;
-        self.0 ^= self.0 << 17;
-        self.0
-    }
-    /// Uniform `f32` in `[0, 1)`. 24 random bits → IEEE single mantissa.
-    pub fn next_f32(&mut self) -> f32 {
-        (self.next_u64() >> 40) as f32 / ((1u64 << 24) as f32)
-    }
-}
-
-/// Sample one token from BF16 logits with optional temperature, top-K,
-/// and top-P (nucleus) filters. `temperature <= 0` falls through to
-/// greedy argmax (the deterministic, reproducible default — same as
-/// [`argmax_bf16_logits`]). `top_k == 0` means "no top-K cap" (full
-/// vocab); `top_p >= 1.0` means "no nucleus truncation".
-///
-/// Greedy is bit-identical to argmax_bf16_logits (same iteration order,
-/// no temperature scaling).
-pub fn sample_bf16_logits(
-    logits_bytes: &[u8],
-    temperature: f32,
-    top_k: usize,
-    top_p: f32,
-    rng: &mut XorshiftRng,
-) -> u32 {
-    if temperature <= 0.0 || top_k == 1 {
-        return argmax_bf16_logits(logits_bytes);
-    }
-    let logits = bf16_bytes_to_f32(logits_bytes);
-    let inv_t = 1.0 / temperature;
-
-    // Top-K: pick the K highest-logit indices, descending. For top_k==0
-    // sort the entire vocab (slow but only paid when sampling — vocab≈248K
-    // takes ~few ms per token, negligible vs the chained decode).
-    let mut indexed: Vec<(usize, f32)> = logits
-        .iter()
-        .enumerate()
-        .map(|(i, &v)| (i, v * inv_t))
-        .collect();
-    let k = if top_k == 0 || top_k > indexed.len() {
-        indexed.len()
-    } else {
-        // Partial sort: select_nth_unstable + sort the head saves the tail.
-        let _ = indexed.select_nth_unstable_by(top_k - 1, |a, b| {
-            b.1.partial_cmp(&a.1).unwrap_or(std::cmp::Ordering::Equal)
-        });
-        top_k
-    };
-    indexed.truncate(k);
-    indexed.sort_unstable_by(|a, b| b.1.partial_cmp(&a.1).unwrap_or(std::cmp::Ordering::Equal));
-
-    // Softmax with max-stabilisation over the kept indices.
-    let max_logit = indexed[0].1;
-    let mut exps: Vec<f32> = indexed.iter().map(|(_, v)| (v - max_logit).exp()).collect();
-    let sum: f32 = exps.iter().sum();
-    if sum <= 0.0 || !sum.is_finite() {
-        return indexed[0].0 as u32;
-    }
-    for e in &mut exps {
-        *e /= sum;
-    }
-
-    // Top-P nucleus: smallest prefix whose cumulative prob ≥ top_p.
-    let nucleus_size = if top_p >= 1.0 {
-        exps.len()
-    } else {
-        let mut cum = 0.0f32;
-        let mut n = exps.len();
-        for (i, &p) in exps.iter().enumerate() {
-            cum += p;
-            if cum >= top_p {
-                n = i + 1;
-                break;
-            }
-        }
-        n.max(1)
-    };
-
-    // Sample from the (renormalised) nucleus.
-    let nucleus_sum: f32 = exps[..nucleus_size].iter().sum();
-    let r: f32 = rng.next_f32() * nucleus_sum;
-    let mut cum = 0.0f32;
-    for i in 0..nucleus_size {
-        cum += exps[i];
-        if cum >= r {
-            return indexed[i].0 as u32;
-        }
-    }
-    indexed[0].0 as u32
-}
-
-/// Greedy argmax over a BF16 logits buffer. Returns the highest-scoring
-/// token id. Used by the runner's first-token smoke path.
-pub fn argmax_bf16_logits(logits_bytes: &[u8]) -> u32 {
-    let logits = bf16_bytes_to_f32(logits_bytes);
-    logits
-        .iter()
-        .enumerate()
-        .fold((0usize, f32::NEG_INFINITY), |(best_i, best_v), (i, &v)| {
-            if v > best_v {
-                (i, v)
-            } else {
-                (best_i, best_v)
-            }
-        })
-        .0 as u32
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -1571,59 +1249,5 @@ mod tests {
             let expect_full = (li + 1) % 4 == 0;
             assert_eq!(is_full_attn_layer(li), expect_full, "layer {li}");
         }
-    }
-
-    #[test]
-    fn bf16_roundtrip_preserves_normal_values() {
-        // PyTorch's BF16 rounding round-trips most "round" F32 values
-        // bit-exactly. Sample a few that have all-zero low mantissa bits
-        // so rounding is a no-op.
-        let bf16_clean = [0.0f32, 1.0, 2.0, 0.5, -3.25, 1024.0];
-        for &v in &bf16_clean {
-            let bytes = f32_to_bf16_bytes(&[v]);
-            let roundtrip = bf16_bytes_to_f32(&bytes);
-            assert_eq!(roundtrip[0], v, "bf16 roundtrip drift at {v}");
-        }
-    }
-
-    #[test]
-    fn rms_norm_then_lm_head_matches_naive_computation() {
-        // Tiny hand-checked case. RMS-normalised (1, 2, 3) has mean_sq=14/3
-        // → rsqrt=sqrt(3/14). With norm_w=(1,1,1) and the HF `(1+w)` unit
-        // offset, the effective scale per channel is 2. Times lm_head row
-        // [1, 0, 0] yields `1 * sqrt(3/14) * 2` ≈ 0.9258. BF16 rounding
-        // is loose enough that we check within 1e-2.
-        let hidden = 3usize;
-        let vocab = 1usize;
-        let h_bytes = f32_to_bf16_bytes(&[1.0, 2.0, 3.0]);
-        let w_bytes = f32_to_bf16_bytes(&[1.0, 1.0, 1.0]);
-        let lm_bytes = f32_to_bf16_bytes(&[1.0, 0.0, 0.0]);
-        let logits = host_final_norm_lm_head(&h_bytes, &w_bytes, &lm_bytes, hidden, vocab, 0.0);
-        let logit = bf16_bytes_to_f32(&logits)[0];
-        let expected = 2.0 * (3.0f32 / 14.0).sqrt();
-        assert!(
-            (logit - expected).abs() < 1e-2,
-            "logit {logit} far from expected {expected}"
-        );
-    }
-
-    #[test]
-    fn rms_norm_unit_offset_zero_weight_is_identity_scale() {
-        // With norm_w = zeros, the HF `(1+w)` convention degrades to
-        // a plain RMSnorm (effective scale = 1). Locks the offset in
-        // semantics — same input as the test above with weight=1 should
-        // give exactly half the logit when weight=0.
-        let hidden = 3usize;
-        let vocab = 1usize;
-        let h_bytes = f32_to_bf16_bytes(&[1.0, 2.0, 3.0]);
-        let w_bytes = f32_to_bf16_bytes(&[0.0, 0.0, 0.0]);
-        let lm_bytes = f32_to_bf16_bytes(&[1.0, 0.0, 0.0]);
-        let logits = host_final_norm_lm_head(&h_bytes, &w_bytes, &lm_bytes, hidden, vocab, 0.0);
-        let logit = bf16_bytes_to_f32(&logits)[0];
-        let expected = (3.0f32 / 14.0).sqrt();
-        assert!(
-            (logit - expected).abs() < 1e-2,
-            "logit {logit} far from expected {expected}"
-        );
     }
 }
