+++ crates/runner/src/qwen36_moe_spec_verify.rs
@@ -3,11 +3,11 @@ use gpu_hal::{GpuBuffer, ScalarType};
 use model_store::BakedStore;
 
 use crate::qwen36_moe_decode::{
-    argmax_bf16_logits, run_chained_decode_fast, DecodeOutputs, LayerBuffers, MtpLayerBuffers,
-    MultiLayerGeom,
+    run_chained_decode_fast, DecodeOutputs, LayerBuffers, MtpLayerBuffers, MultiLayerGeom,
 };
 use crate::qwen36_moe_host::lookup_embed_row;
 use crate::qwen36_moe_lm_head::{launch_lm_head_from_final_hidden_bytes, LmHeadBuffers};
+use crate::qwen36_moe_logits::argmax_bf16_logits;
 use crate::qwen36_moe_loop::Qwen36DecodeLoopState;
 use crate::qwen36_moe_mtp::{MtpChainScratch, MtpForwardScratch};
 use crate::qwen36_moe_persistent_decode::PersistentScratch;
